svc-workflows/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
httpx>=0.27
aiohttp>=3.10
redis>=5.0
langchain-core>=0.3
langchain-openai>=0.3
langchain-anthropic>=0.3
langchain-google-genai>=2.0
tavily-python>=0.5
